sys: add script to configure Inotify watches limit
For an intelligent IDE, it is essential to be in the know about any
external changes in files it working with - e.g. changes made by VCS, or
build tools, or code generators, etc. For that reason, the IntelliJ
platform spins a background process to monitor such changes. The method
it uses is platform-specific, and on Linux, it is the Inotify facility.

Inotify requires a "watch handle" to be set for each directory in the
project. Unfortunately, the default limit of watch handles may not be
enough for reasonably sized projects, and reaching the limit will force
IntelliJ platform to fall back to recursive scans of directory trees.

Learn more → https://youtrack.jetbrains.com/articles/IDEA-A-2/Inotify-Watches-Limit